Velvet: Add tests to velvet to check areas trees behaviour, aggregates, tally sheets (#586)
Main Issue: https://github.com/sequentech/meta/issues/1378

---------

+++ packages/velvet/src/fixtures/fixtures.rs
@@ -1,10 +1,12 @@
 // SPDX-FileCopyrightText: 2023 Kevin Nguyen <kevin@sequentech.io>
+// SPDX-FileCopyrightText: 2024 Eduardo Robles <edu@sequentech.io>
 //
 // SPDX-License-Identifier: AGPL-3.0-only
 
 use anyhow::Result;
 use sequent_core::ballot::Contest;
 use std::collections::HashMap;
+use std::env;
 use std::fs;
 use std::io::Write;
 use std::path::PathBuf;
@@ -58,6 +60,26 @@ impl TestFixture {
         areas: Vec<Uuid>,
     ) -> Result<ElectionConfig> {
         let election = super::elections::get_election_config_1(election_event_id, areas);
+
+        let mut path = self
+            .input_dir_configs
+            .join(format!("election__{}", election.id));
+        fs::create_dir_all(path.as_path())?;
+
+        path.push("election-config.json");
+        let mut file = fs::File::create(path)?;
+        writeln!(file, "{}", serde_json::to_string(&election)?)?;
+
+        Ok(election)
+    }
+
+    #[instrument]
+    pub fn create_election_config_2(
+        &self,
+        election_event_id: &Uuid,
+        areas: Vec<(Uuid, Option<Uuid>)>,
+    ) -> Result<ElectionConfig> {
+        let election = super::elections::get_election_config_3(election_event_id, areas);
 
         let mut path = self
             .input_dir_configs
@@ -165,8 +187,10 @@ impl TestFixture {
 
 impl Drop for TestFixture {
     fn drop(&mut self) {
-        fs::remove_file(&self.config_path).unwrap();
-        fs::remove_dir_all(&self.root_dir).unwrap();
+        if env::var("CLEANUP_FILES").unwrap_or("true".to_string()) == "true" {
+            fs::remove_file(&self.config_path).unwrap();
+            fs::remove_dir_all(&self.root_dir).unwrap();
+        }
     }
 }
